requirements.txt
apache-airflow==2.9.*
beautifulsoup4==4.11.*
pathvalidate==3.2.*
pandas==2.2.*
openpyxl==3.1.*
lxml==5.2.*
clickhouse-sqlalchemy==0.3.*
clickhouse-connect==0.7.*
pandahouse==0.2.*
